moved GUI to seperate file
